implementing serialization for IBaseSynapses

+++ split_synapse_handler/Exp2PostSynapse.cpp
@@ -29,8 +29,8 @@ Exp2PostSynapse::Exp2PostSynapse(
 }
 
 Exp2PostSynapse::Exp2PostSynapse(
-		const unsigned long id,
-		const unsigned long presynapse_id,
+		const unsigned long long id,
+		const unsigned long long presynapse_id,
 		const number& location,
 		const number& tau1,
 		const number& tau2,
