+++ split_synapse_handler/Exp2PreSynapse.h
@@ -30,8 +30,8 @@ public:
 			const number& onset);
 
 	Exp2PreSynapse(
-			const unsigned long id,
-			const unsigned long postsynapse_id,
+			const unsigned long long id,
+			const unsigned long long postsynapse_id,
 			const number& location,
 			const number& onset);
