+++ split_synapse_handler/IPreSynapse.cpp
@@ -12,8 +12,8 @@ namespace cable_neuron {
 namespace synapse_handler {
 
 IPreSynapse::IPreSynapse(
-		const unsigned long id,
-		const unsigned long postsynapse_id,
+		const unsigned long long id,
+		const unsigned long long postsynapse_id,
 		const number& location)
 
 :m_id(id),
